change init

+++ app/helpers/application_helper.rb
@@ -11,7 +11,7 @@ module ApplicationHelper
   end
 
   def time_mini(start_time, end_time)
-    (end_time.to_f - start_time.to_f)
+    (end_time.to_i - start_time.to_i)
   end
   
 end
